Keyword Detection for Identifying Cancer Tumor Boards
This project involves developing an algorithm to determine whether an HTML page contains information about cancer tumor boards. Tumor boards are consiliums of doctors from various disciplines discussing cancer cases. The dataset includes training and test CSV files, corresponding HTML documents, and a file of keywords related to tumor board types. The aim is to build a predictive model that classifies the pages based on their content, along with bonus tasks to identify tumor board types and schedules. The final output will be a CSV file with predictions for the test data.